requirements.txt
fastapi
uvicorn
python-multipart
requests
pypdf
faiss-cpu
sentence-transformers
numpy